pyproject.toml
[project]
name = "neoxtractor"
version = "0.1.0"
description = "NeoX NXPK / EXPK (.npk) extraction and viewer tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arc4>=0.4.0",
    "astc-encoder-py>=0.1.11",
    "bitstring>=4.3.1",
    "lz4>=4.4.4",
    "numpy>=2.2.5",
    "p>=1.5.0",
    "pillow>=11.2.1",
    "pycryptodome>=3.22.0",
    "pymeshio>=3.0.1",
    "pyside6>=6.9.0",
    "zstandard>=0.23.0",
]

[tool.uv]
no-binary-package = ["astc-encoder-py"]